Changed the i2c address variabel to uin8_t. This fixes #1. Thanks @dzalf for reporting this.

+++ src/TMP1075.h
@@ -84,7 +84,7 @@ namespace TMP1075 {
       float convertToCelsius(const int16_t value);
       int16_t convertFromCelsius(const float value);
       uint16_t readRegister(const uint8_t reg);
-      uint16_t i2cAdress;
+      uint8_t i2cAdress;
       uint8_t configRegister;
   };
 }
